Fix single client mode
I noticed that the if statement that handled if the client was in single client mode or not was a bit messed up and never actually checked if the client was in single client mode. (This was after trying to use it and having it not work) Changing the elif statement to just else ended up allowing it to work for me.